requirements.txt
fastapi==0.115.0
pandas==2.2.3
uvicorn==0.30.6
python-multipart==0.0.10
requests
pyjwt
passlib